Check if namepsace is alive during list.
In contrast to the netns handling in "ip netns" we bind mount the
/proc/<pid>/ns directory. This gives use the possibility to check
whether the namespace is still alive or not. It the "init" process in
the namespace dies, the bind mount gets invalid and the pid file in that
directory is not accessible anymore.

If we find such dead namespace during list, we remove it automatically,
because it is not usable anymore.